requirements.txt
attrs==21.4.0
Automat==20.2.0
black==21.12b0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
mypy-extensions==0.4.3
pathspec==0.9.0
peewee==3.14.8
peewee-migrations==0.3.30
platformdirs==2.4.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.12.0
pyOpenSSL==21.0.0
python-dotenv==0.19.2
service-identity==21.1.0
six==1.16.0
tomli==1.2.3
Twisted==21.7.0
typing_extensions==4.0.1
zope.interface==5.4.0
